third_party/blink/web_tests/fast/dom/shadow/title-element-in-shadow-expected.txt

This test ensures that title elements in a shadow subtree do not affect document.title attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

create a title element and insert it to document.
PASS document.title is "document title"
remove title element from document.
PASS document.title is ""
create a shadow root whose host is already in document, create a subtree which contains a title element, and add the subtree to the shadow root.
PASS document.title is ""
remove the subtree from document.
PASS document.title is ""
create a subtree, add a shadow root which contains a title element to the subtree, and insert the subtree to document.
PASS document.title is ""
remove the subtree from document
PASS document.title is ""
PASS successfullyParsed is true

TEST COMPLETE
